Fix type-safe formatting warnings
Clean up printf/debug formatting across DMX, network, HTTP, and showfile code by using explicit casts and clearer control flow where needed. The change also adds a missing standard header, applies a few `[[nodiscard]]` and signature cleanups, and updates the remote-config datetime code to use the current global UTC offset helpers. Overall this looks like a warning-reduction and consistency pass without intended behavior changes.

+++ lib-network/src/emac/gd32/emac_debug.cpp
@@ -39,6 +39,6 @@ void emac_debug_run() {
 #endif
 
     if ((rxfifo_drop != 0) || (rxdma_drop != 0)) {
-        printf("%u: RxFIFO: %u RxDMA: %u\n", ++s_counter, rxfifo_drop, rxdma_drop);
+        printf("%u: RxFIFO: %u RxDMA: %u\n", static_cast<unsigned>(++s_counter), static_cast<unsigned>(rxfifo_drop), static_cast<unsigned>(rxdma_drop));
     }
 }
